Moved get_seasonality from evaluation to time_feature.
