+++ src/gluonts/model/n_beats/_network.py
@@ -20,8 +20,8 @@ import numpy as np
 
 # First-party imports
 from gluonts.core.component import validated
-from gluonts.evaluation import get_seasonality
 from gluonts.model.common import Tensor
+from gluonts.time_feature import get_seasonality
 
 VALID_N_BEATS_STACK_TYPES = "G", "S", "T"
 VALID_LOSS_FUNCTIONS = "sMAPE", "MASE", "MAPE"
