+++ src/gluonts/model/naive_2/_predictor.py
@@ -23,10 +23,10 @@ import statsmodels.api as sm
 # First-party imports
 from gluonts.core.component import validated
 from gluonts.dataset.common import DataEntry
-from gluonts.evaluation import get_seasonality
 from gluonts.model.forecast import Forecast, SampleForecast
 from gluonts.model.predictor import RepresentablePredictor
 from gluonts.support.pandas import forecast_start
+from gluonts.time_feature import get_seasonality
 
 
 def seasonality_test(past_ts_data: np.array, season_length: int) -> bool:
