+++ src/gluonts/model/wavenet/_estimator.py
@@ -34,7 +34,10 @@ from gluonts.support.util import (
     copy_parameters,
     get_hybrid_forward_input_names,
 )
-from gluonts.time_feature import time_features_from_frequency_str
+from gluonts.time_feature import (
+    get_seasonality,
+    time_features_from_frequency_str,
+)
 from gluonts.transform import (
     AddAgeFeature,
     AddObservedValuesIndicator,
@@ -87,21 +90,6 @@ class QuantizeScaled(SimpleTransformation):
         )
         data[self.scale] = np.array([scale])
         return data
-
-
-def _get_seasonality(freq: str, seasonality_dict: Dict) -> int:
-    match = re.match(r"(\d*)(\w+)", freq)
-    assert match, "Cannot match freq regex"
-    multiple, base_freq = match.groups()
-    multiple = int(multiple) if multiple else 1
-    seasonality = seasonality_dict[base_freq]
-    if seasonality % multiple != 0:
-        logging.warning(
-            f"multiple {multiple} does not divide base seasonality {seasonality}."
-            f"Falling back to seasonality 1"
-        )
-        return 1
-    return seasonality // multiple
 
 
 class WaveNetEstimator(GluonEstimator):
@@ -221,7 +209,7 @@ class WaveNetEstimator(GluonEstimator):
         self.num_parallel_samples = num_parallel_samples
 
         seasonality = (
-            _get_seasonality(
+            get_seasonality(
                 self.freq,
                 {
                     "H": 7 * 24,
