+++ src/gluonts/time_feature/__init__.py
@@ -25,6 +25,7 @@ from ._base import (
 )
 from .holiday import SPECIAL_DATE_FEATURES, SpecialDateFeatureSet
 from .lag import get_lags_for_frequency
+from .seasonality import get_seasonality
 
 __all__ = [
     "DayOfMonth",
@@ -39,6 +40,7 @@ __all__ = [
     "SpecialDateFeatureSet",
     "get_lags_for_frequency",
     "time_features_from_frequency_str",
+    "get_seasonality",
 ]
 
 # fix Sphinx issues, see https://bit.ly/2K2eptM
